move sendStateSnapshotTrace() to dialog manager
